back/requirements.txt
Django
django-allauth
djoser
djangorestframework
django-compressor
django-sass-processor
drf-yasg
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-jaeger
opentelemetry-instrumentation
opentelemetry-instrumentation-django
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-requests
requests
django-prometheus
django-plotly-dash
django-filter
django-cors-headers
pandas
asgiref
autopep8
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pycodestyle
PyJWT
pytz
sqlparse
toml
tzdata
django-taggit
djangorestframework-simplejwt
django-jazzmin
django-import-export
django-crispy-forms
django-mathfilters
django-ckeditor
psycopg2-binary
django-pglocks
django-ckeditor-5
drf-excel
gunicorn
gevent 
traceback